CV checker/requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
pyngrok==1.2.0
python-dotenv==1.0.0
groq==0.4.0
gunicorn==22.0.0




